requirements.txt
africastalking==1.2.4
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
contextlib2==21.6.0
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.3
python-dotenv==0.21.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
schema==0.7.5
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
